+++ packages/svelte-undp-components/src/lib/components/maplibre/hillshade/index.ts
@@ -3,11 +3,13 @@ import HillshadeExaggeration from './HillshadeExaggeration.svelte';
 import HillshadeHighlightColor from './HillshadeHighlightColor.svelte';
 import HillshadeIlluminationDirection from './HillshadeIlluminationDirection.svelte';
 import HillshadeShadowColor from './HillshadeShadowColor.svelte';
+import HillshadeMethod from './HillshadeMethod.svelte';
 
 export {
 	HillshadeAccentColor,
 	HillshadeExaggeration,
 	HillshadeHighlightColor,
 	HillshadeIlluminationDirection,
-	HillshadeShadowColor
+	HillshadeShadowColor,
+	HillshadeMethod
 };